Parsing fleshed out a bit
I think I have the parsing down a bit more:
    MParser has a stack
        Each stack entry is a list of the previous positions before each
action was taken
        There's a macro to wrap parsing each atom; Each time you call it
it pushes onto the stack, and if it fails it pops off and returns an Err

It's definitely not done yet, but it's on its way. I probably need to
just start writing with it and parsing a few atoms to see how it ends up
feeling.