Move parsing into the self, default DFS method
Everything is so much nicer now. Got rid of the generic constraint on
the trait so now I can have a default implementation of self.parse().
That method handles the looping and parsing children, now each atom just
needs to implement parse_self and optionally parse_child and
get_children if it has any children.